COMP: RunLength signed / unsigned comparison
To address:

 c:\jenkins\workspace\itkgerritwindows\itk-src\modules\remote\texturefeatures\include\itkRunLengthTextureFeaturesImageFilter.hxx(404):
 warning C4018: '<': signed/unsigned mismatch

+++ Modules/Filtering/TextureFeatures/include/itkRunLengthTextureFeaturesImageFilter.hxx
@@ -387,10 +387,10 @@ RunLengthTextureFeaturesImageFilter<TInputImage, TOutputImage>::IncreaseHistogra
   }
   offsetDistance = std::sqrt(offsetDistance);
 
-  int offsetDistanceBin = (int)((offsetDistance * pixelDistance - m_MinDistance) /
-                                ((m_MaxDistance - m_MinDistance) / (float)m_NumberOfBinsPerAxis));
+  int offsetDistanceBin = static_cast<int>((offsetDistance * pixelDistance - m_MinDistance) /
+                                           ((m_MaxDistance - m_MinDistance) / (float)m_NumberOfBinsPerAxis));
 
-  if (offsetDistanceBin < m_NumberOfBinsPerAxis)
+  if (offsetDistanceBin < static_cast<int>(m_NumberOfBinsPerAxis))
   {
     totalNumberOfRuns++;
     hist[curentInNeighborhoodPixelIntensity][offsetDistanceBin]++;
